requirements.txt
aiohttp==3.6.2
aiomysql==0.0.21
alembic==1.4.3
amqp==2.6.1
aniso8601==8.1.0
anyjson==0.3.3
asgiref==3.2.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
bcrypt==3.2.0
billiard==3.6.3.0
celery==4.4.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Django==3.1.4
djangorestframework==3.11.1
drf-extensions==0.6.0
idna==2.10
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
mysql-connector==2.2.9
mysqlclient==2.0.1
numpy==1.19.4
pandas==1.1.5
paramiko==2.7.2
psutil==5.7.3
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.4.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==1.3.0
Werkzeug==1.0.1
yarl==1.6.3
zipp==3.4.0
